+++ drift_dev/lib/src/writer/writer.dart
@@ -98,14 +98,34 @@ abstract class _NodeOrWriter {
   /// Returns a Dart expression evaluating to the [converter].
   AnnotatedDartCode readConverter(AppliedTypeConverter converter,
       {bool forNullable = false}) {
-    final fieldName = forNullable && converter.canBeSkippedForNulls
-        ? converter.nullableFieldName
-        : converter.fieldName;
+    if (converter.owningColumn == null) {
+      // Type converters applied to individual columns in the result set of a
+      // `SELECT` query don't have an owning table. We instead write the
+      // expression here.
+      return AnnotatedDartCode.build((b) {
+        final implicitlyNullable =
+            converter.canBeSkippedForNulls && forNullable;
 
-    return AnnotatedDartCode([
-      ...entityInfoType(converter.owningColumn.owner).elements,
-      '.$fieldName',
-    ]);
+        if (implicitlyNullable) {
+          b.addSymbol('NullAwareTypeConverter.wrap(', AnnotatedDartCode.drift);
+        }
+
+        b.addCode(converter.expression);
+
+        if (implicitlyNullable) {
+          b.addText(')');
+        }
+      });
+    } else {
+      final fieldName = forNullable && converter.canBeSkippedForNulls
+          ? converter.nullableFieldName
+          : converter.fieldName;
+
+      return AnnotatedDartCode([
+        ...entityInfoType(converter.owningColumn!.owner).elements,
+        '.$fieldName',
+      ]);
+    }
   }
 
   /// A suitable typename to store an instance of the type converter used here.